requirements.txt
amqp==2.2.2
Babel==2.5.0
billiard==3.5.0.3
celery==4.1.0
Cerberus==1.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
django-annoying==0.10.4
django-cors-headers==2.1.0
django-filter==1.0.4
django-material==1.0.0
django-suit-redactor==0.0.4
django-tinymce==2.7.0
django-tinymce4-lite==1.7.1
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
flower==0.9.2
futures==3.1.1
gunicorn==19.7.1
idna==2.6
jdatetime==1.9.0
jsmin==2.2.2
kombu==4.1.0
Markdown==2.6.8
olefile==0.44
pika==0.12.0
Pillow==4.2.1
psycopg2==2.7.3.2
pyfcm==1.4.3
PyJWT==1.5.2
pymongo==3.7.2
python-dateutil==2.6.1
python-memcached==1.58
pytz==2017.2
raven==6.2.1
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
tornado==4.5.2
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
XlsxWriter==1.0.5
xlwt==1.3.0
yuicompressor==2.4.8
